# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Patient:
    description: Patient information
    properties:
      amka:
        description: '@Description The patient''s AMKA (Social Security Number)'
        example: "12345678901"
        type: string
      created_at:
        description: '@Description The time when the record was created'
        example: "2023-01-01T00:00:00Z"
        type: string
      deleted_at:
        description: '@Description The time when the record was deleted (soft delete)'
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        description: '@Description The unique identifier for the record'
        example: 1
        type: integer
      registration:
        description: '@Description The date of patient registration'
        example: "2023-01-01T00:00:00Z"
        type: string
      updated_at:
        description: '@Description The time when the record was last updated'
        example: "2023-01-01T00:00:00Z"
        type: string
      user_id:
        description: '@Description The ID of the associated user'
        example: 1
        type: integer
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clinic management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Clinic Management API
  version: "1.0"
paths:
...
  /patients:
    get:
      consumes:
      - application/json
      description: Get a list of all patients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Patient'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List patients
      tags:
      - patients
    post:
      consumes:
      - application/json
      description: Add a new patient to the system
      parameters:
      - description: Patient data
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/models.Patient'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new patient
      tags:
      - patients
  /patients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a patient by their ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a patient
      tags:
      - patients
    get:
      consumes:
      - application/json
      description: Get a patient's details by their ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Patient'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a patient
      tags:
      - patients
    put:
      consumes:
      - application/json
      description: Update a patient's details
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated patient object
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/models.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a patient
      tags:
      - patients
  /patients/bulk:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple patients using a CSV file
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Bulk upload patients
      tags:
      - patients
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"